Add meta content rendering for hero video
Refined the HTML comment for poster image and added template PHP for rendering meta content in the browser. The new code ensures that if there are either custom video or poster images available, an accompanying section with supportive information like logo, title, description and contact number will be presented as well.